Add methods for QStandardPathLocation.
Also refactored QStandardPathLocation emum to separate module in qtcore.
