+++ qttypes/src/qtcore/mod.rs
@@ -2,6 +2,7 @@ mod primitives;
 mod qbytearray;
 mod qchar;
 mod qlist;
+mod qstandardpaths;
 mod qstring;
 mod qurl;
 
@@ -9,5 +10,6 @@ pub use self::primitives::qreal;
 pub use self::qbytearray::QByteArray;
 pub use self::qchar::UnicodeVersion;
 pub use self::qlist::{QListIterator, QStringList, QVariantList};
+pub use self::qstandardpaths::QStandardPathLocation;
 pub use self::qstring::{NormalizationForm, QString};
 pub use self::qurl::QUrl;
